change adder amount to come from zone

+++ src/main/resources/db/migration/procedures/brs/action_functions/R__save_postal_code_zone_field_to_project.sql
@@ -6,7 +6,7 @@ AS
 $function$
 declare
   v_metro_area_id int;
-  v_adder_amount int;
+  v_adder_amount numeric;
 BEGIN
   -- we dont need the ppse_id yet but it is easier to add now vs later if we need it so i put it in
